Extract Processor
