+++ src/Microsoft.Extensions.Logging.Console/Internal/LogMessageEntry.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+
+namespace Microsoft.Extensions.Logging.Console.Internal
+{
+    public struct LogMessageEntry
+    {
+        public string LevelString;
+        public ConsoleColor? LevelBackground;
+        public ConsoleColor? LevelForeground;
+        public ConsoleColor? MessageColor;
+        public string Message;
+    }
+}
